requirements.txt
fastapi==0.73.0
python-dotenv==0.1.5
pydantic==1.9.0
deta==1.1.0
